backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9  # PostgreSQL adapter for Supabase

# Task Queue
celery==5.3.6
redis==5.0.1

# Video Processing
ffmpeg-python==0.2.0
pillow==10.2.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.0  # For LLM API calls (OpenRouter)
requests==2.31.0  # For testing endpoints

# Monitoring & Logging
python-json-logger==2.0.7

# WebSockets (for real-time updates)
websockets==12.0

# Validation
python-magic==0.4.27  # MIME type detection

# Transcription & Speech Processing
faster-whisper # Faster Whisper implementation (open-source, free)
# torch - Install CPU-only FIRST: pip install torch torchaudio --index-url https://download.pytorch.org/whl/cpu
# Then install silero-vad which will use the existing CPU torch

# Audio Analysis
librosa
soundfile
# silero-vad - Install AFTER CPU torch to avoid CUDA dependencies
pydub  # Audio manipulation

# Scene Detection
scenedetect[opencv] # PySceneDetect for scene change detection

# Face Detection & Computer Vision
mediapipe # Google's face detection (free, fast)

# LLM Integration (Free options - HuggingFace)
transformers  # For using HuggingFace models
# Note: torch is already listed above

# Optional: OpenAI (if you have API key and want to use it)
# openai==1.6.1

# Note: ffmpeg is installed as a system package via nixPkgs